+++ email_tool/backend/models/__init__.py
@@ -1,0 +1,17 @@
+from .base import Base
+from .campaign import Campaign
+from .template import Template
+from .placeholder import Placeholder
+from .localized_copy import LocalizedCopy
+from .generated_email import GeneratedEmail
+from .playwright_result import PlaywrightResult
+
+__all__ = [
+    'Base',
+    'Campaign',
+    'Template',
+    'Placeholder',
+    'LocalizedCopy',
+    'GeneratedEmail',
+    'PlaywrightResult',
+]
