+++ email_tool/backend/models/campaign.py
@@ -1,0 +1,15 @@
+from sqlalchemy import Column, Integer, String, DateTime
+from sqlalchemy.orm import relationship
+from datetime import datetime
+from .base import Base
+
+class Campaign(Base):
+    __tablename__ = 'campaign'
+
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String, nullable=False)
+    created_at = Column(DateTime, default=datetime.utcnow)
+
+    templates = relationship('Template', back_populates='campaign')
+    copies = relationship('LocalizedCopy', back_populates='campaign')
+    generated_emails = relationship('GeneratedEmail', back_populates='campaign')
