+++ email_tool/backend/models/template.py
@@ -1,0 +1,16 @@
+from sqlalchemy import Column, Integer, String, Text, DateTime, ForeignKey
+from sqlalchemy.orm import relationship
+from datetime import datetime
+from .base import Base
+
+class Template(Base):
+    __tablename__ = 'template'
+
+    id = Column(Integer, primary_key=True, index=True)
+    campaign_id = Column(Integer, ForeignKey('campaign.id'))
+    filename = Column(String, nullable=False)
+    content = Column(Text, nullable=False)
+    created_at = Column(DateTime, default=datetime.utcnow)
+
+    campaign = relationship('Campaign', back_populates='templates')
+    placeholders = relationship('Placeholder', back_populates='template')
